+++ examples/data-sources/iosxe_bgp_l2vpn_neighbor/data-source.tf
@@ -1,5 +1,0 @@
-data "iosxe_bgp_l2vpn_neighbor" "example" {
-  asn     = "65000"
-  af_name = "evpn"
-  ip      = "3.3.3.3"
-}