+++ examples/resources/iosxe_bgp_l2vpn_evpn_neighbor/resource.tf
@@ -1,6 +1,5 @@
-resource "iosxe_bgp_l2vpn_neighbor" "example" {
+resource "iosxe_bgp_l2vpn_evpn_neighbor" "example" {
   asn                    = "65000"
-  af_name                = "evpn"
   ip                     = "3.3.3.3"
   activate               = true
   send_community         = "both"