+++ examples/resources/iosxe_bgp_l2vpn_neighbor/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxe_bgp_l2vpn_neighbor.example "Cisco-IOS-XE-native:native/router/Cisco-IOS-XE-bgp:bgp=65000/address-family/no-vrf/l2vpn=evpn/l2vpn-evpn/neighbor=3.3.3.3"